--- src/components/TestScreen.style.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components/native';
-
-export const S = {
-  TestScreenContainer: styled.View`
-    flex: 1;
-  `,
-};
