--- src/mocks/MyPageMocks.ts
@@ -1,30 +1,0 @@
-import { MyReservation } from '@/types/MyPageScreen';
-
-export const myReservation: MyReservation[] = [
-  {
-    reservationId: '23',
-    popupId: 1,
-    popupName: '블랙핑크 팝업스토어asdfasdfasdfasdfasdf',
-    reservationDate: '2025-05-30',
-    reservationTime: '11:00',
-    reservationDay: 'MON',
-    address: '서울 영등포구 여의대로 108, 더현대서울asdfasdfadsfasdf',
-    latitude: 37.1234561234567,
-    longitude: 127.37123456123456,
-    qrImage:
-      'iVBORw0KGgoAAAANSUhEUgAAAPoAAAD6AQAAAACgl2eQAAABdUlEQVR4Xu2XTaoEIRCDCzyAR/LqHskDCL4kZQ/TLmb9AoamccqvF6F+dGL9Vo8zcugCqQukLpC6QOr/ADOoNRoCs/bAA1U3gOs+o0W0MgLME/QCaG2NKAjUhTWMmwKwCaAMvn2BzBfWtgCXU5WW5BO0AtT+AD7P2f4OwEfafUWMANqUNSwwkNlBbB8zACrEGuG6MI2Rqe+SswDmHr8tzS6eKftDLwA2Cw8UrHGyMFOvbFoAaU2Vhl02/thZdAKq6o0M06SS23PACCAjpwwwjnfPuBOgTpE1PotH5KvkPACd7GUPYcJL57sXIHcsPJKhq69qzw5YmmCot2AH0Sw3zAAVGCcws0az75IzAXpe1xGc/EvIIaY54ASkWHUayMEF314Af/KuyF12TaXl/MoJUOMv5qtpIMuvH7CtaQ40ZaofNWkEaC3XKjxPAJniNNbPc4g5AFwCyK7hrm7vZsDT/gzq9l6OK4oD8EMXSF0gdYHUBVIewB+eKbgofoNSbQAAAABJRU5ErkJggg==',
-  },
-  {
-    reservationId: 24,
-    popupId: 2,
-    popupName: '뉴진스 팝업스토어',
-    reservationDate: '2025-05-26',
-    reservationTime: '15:00',
-    reservationDay: 'FRI',
-    address: '서울 강남구 테헤란로 123, 스타팝업존',
-    latitude: 37.501234,
-    longitude: 127.039123,
-    qrImage:
-      'iVBORw0KGgoAAAANSUhEUgAAAPoAAAD6AQAAAACgl2eQAAABbUlEQVR4Xu2YzUoDMRCG3zVUlVg7/ScTKzSkVUJWwlZhcAGtgVuAFrAVFQVKZ8qRhGxs/VtxcLsjDLz4+Tn8fn8M1QmSTIAJd8AJd8AJd+AGwAp8FhNIVUCZ1jqQWeDwRB9w4At7EomEIeqQqL3BAMRYsCzDtRjZgEdAFUvF+f7ZMRHOQR8zD3BmUOeUED7BbXEQf4Yo0ODDHu7aw7m8LS8nhkhnobKxCnaxk8pykzy09CR1nSRHiyQt+a0rQs1MTniLlD7sZpKX3TAZMRkbHZIQXcawM5tI+Eb6MpPaCXgEO3wxNKAljxTFloypN8AFqVrKPw/FFcB2YMLF9Yey0z0FZDCJrboPoJQAAAABJRU5ErkJggg==',
-  },
-];
