--- src/mocks/PopUpDetailItemMocks.ts
@@ -1,130 +1,0 @@
-import { ItemPathType, ItemUrlType } from '@/types/DetailScreen';
-
-export const ItemMocks: ItemPathType[] = [
-  {
-    itemId: 1,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] MOON ...',
-    price: 14800,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 2,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] 2024 ...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 3,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 4,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 1,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] MOON ...',
-    price: 14800,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 2,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] 2024 ...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 3,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 4,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 1,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] MOON ...',
-    price: 14800,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 2,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] 2024 ...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 3,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 4,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-];
-
-export const HotItemMocks: ItemPathType[] = [
-  {
-    itemId: 1,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] MOON ...',
-    price: 14800,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 2,
-    title: '[BLACKPINK THE GAME COUPON CARD COLLECTION] 2024 ...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 3,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 4,
-    title: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imagePath: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-];
-
-export const EntireItemMocks: ItemUrlType[] = [
-  {
-    itemId: 1,
-    name: '[BLACKPINK THE GAME COUPON CARD COLLECTION] MOON ...',
-    price: 14800,
-    imageUrl: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 2,
-    name: '[BLACKPINK THE GAME COUPON CARD COLLECTION] 2024 ...',
-    price: 22000,
-    imageUrl: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 3,
-    name: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imageUrl: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-  {
-    itemId: 4,
-    name: '[WELCOMEBACKENCODE] 2NE WIND BREAKER...',
-    price: 22000,
-    imageUrl: 'https://image.aladin.co.kr/product/17920/59/cover500/k142534034_1.jpg',
-  },
-];
