--- src/mocks/SurveyQuestionsMocks.ts
@@ -1,44 +1,0 @@
-import { SurveyQuestionsResponse } from '@/types/api/ApiResponseType';
-
-export const SurveyQuestionsMock: SurveyQuestionsResponse = [
-  {
-    surveyId: 25,
-    options: [
-      { choiceId: 1, content: '보기1' },
-      { choiceId: 2, content: '보기2' },
-      { choiceId: 3, content: '보기3' },
-      { choiceId: 4, content: '보기4' },
-      { choiceId: 5, content: '보기5' },
-    ],
-  },
-  {
-    surveyId: 26,
-    options: [
-      { choiceId: 6, content: '보기1' },
-      { choiceId: 7, content: '보기2' },
-      { choiceId: 8, content: '보기3' },
-      { choiceId: 9, content: '보기4' },
-      { choiceId: 10, content: '보기5' },
-    ],
-  },
-  {
-    surveyId: 27,
-    options: [
-      { choiceId: 11, content: '보기1' },
-      { choiceId: 12, content: '보기2' },
-      { choiceId: 13, content: '보기3' },
-      { choiceId: 14, content: '보기4' },
-      { choiceId: 15, content: '보기5' },
-    ],
-  },
-  {
-    surveyId: 28,
-    options: [
-      { choiceId: 16, content: '보기1' },
-      { choiceId: 17, content: '보기2' },
-      { choiceId: 18, content: '보기3' },
-      { choiceId: 19, content: '보기4' },
-      { choiceId: 20, content: '보기5' },
-    ],
-  },
-];
