+++ src/types/PopUpEntryScreen.ts
@@ -1,13 +1,13 @@
 export type RecommendedItem = {
   itemId: number;
   title: string;
-  imagePath: number; // 추후 string으로 변경
+  imagePath: number;
   price: number;
 };
 
 export type PopularItem = {
   itemId: number;
   title: string;
-  imagePath: number; // 추후 string으로 변경
+  imagePath: number;
   price: number;
 };
